Visual Studio Static Analysis fixes for OT
I have used the build-in visual studio static analysis tool
on the OT project and fixed all the results.

In the process I've rewrote OT_ASSERT, and created a new macro
OT_FAIL.  These macros now insert std::terminate, giving a much
stronger assurance that the execution will not continue.

I've also fixed many warnings from the normal compile.

Overall, I hope that OT is more sold from these changes.

Cam.

+++ include/otlib/OTCredential.h
@@ -696,7 +696,7 @@ public:
     EXPORT bool LoadSubkeyFromString (const OTString & strInput, const OTString & strSubID, OTPassword * pImportPassword=NULL);
     EXPORT bool LoadSubcredentialFromString(const OTString & strInput, const OTString & strSubID, OTPassword * pImportPassword=NULL);
     // ------------------------------
-    EXPORT int   GetSubcredentialCount() const;
+    EXPORT size_t GetSubcredentialCount() const;
     EXPORT const OTSubcredential * GetSubcredential (const OTString & strSubID, const listOfStrings * plistRevokedIDs=NULL) const;
     EXPORT const OTSubcredential * GetSubcredentialByIndex (int nIndex) const;
     EXPORT const std::string GetSubcredentialIDByIndex(int nIndex) const;
