+++ include/otlib/OTSymmetricKey.h
@@ -166,7 +166,7 @@ private:
 	bool      m_bIsGenerated;     // GetKey asserts if this is false; GenerateKey asserts if it's true.
 	bool      m_bHasHashCheck;    // If a hash-check fo the Derived Key has been made yet.
 	// ---------------------------------------------
-	uint16_t  m_nKeySize;         // The size, in bits. For example, 128 bit key, 256 bit key, etc.
+	uint32_t  m_nKeySize;         // The size, in bits. For example, 128 bit key, 256 bit key, etc.
 	// ---------------------------------------------
 	uint32_t  m_uIterationCount;  // Stores the iteration count, which should probably be at least 2000. (Number of iterations used while generating key from passphrase.)
 	// ---------------------------------------------
